[module2-task1] provide webpack.config
split entrypoint into ES6-modules
composed modules by their roles
