+++ src/components/index.js
@@ -1,0 +1,6 @@
+export * from './statistics';
+export * from './navigation';
+export * from './userLevel';
+export * from './sort';
+export * from './films';
+export * from './filmDetailsPopup';
